requirements.txt
openai
python-dotenv
requests
pandas
beautifulsoup4
pytest
deep-translator
pytrends
pillow
